move vcf parsing into ingestion
